+++ beep/beep_darwin.go
@@ -162,7 +162,7 @@ func Init() {
 }
 
 func PlayStart() {
-	if disabled {
+	if disabled.Load() {
 		return
 	}
 	soundOnce.Do(initSound)
@@ -170,7 +170,7 @@ func PlayStart() {
 }
 
 func PlayEnd() {
-	if disabled {
+	if disabled.Load() {
 		return
 	}
 	soundOnce.Do(initSound)
@@ -178,7 +178,7 @@ func PlayEnd() {
 }
 
 func PlayError() {
-	if disabled {
+	if disabled.Load() {
 		return
 	}
 	soundOnce.Do(initSound)
@@ -186,7 +186,7 @@ func PlayError() {
 }
 
 func PlayDenied() {
-	if disabled {
+	if disabled.Load() {
 		return
 	}
 	soundOnce.Do(initSound)